requirements.txt
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=9.0.0
PyYAML>=6.0

